[bflat-release/7.0] Backport of 18eaeb27c0f69d9ebd60fbe9301014d405e40017 (#42)
* Pass NodeFactory to ObjectData's Section (#77967)

This is a preparatory change for something I'm working on, but makes sense on its own - notice several node types that had to hold on to extra state just because we didn't have context in the `Section` property. Got rid of the extra state as well. A couple bytes of memory saved here and there.

* breaks